Move types back
